chore: make helper functions shared
