+++ libs/application/templates/inao/financial-statement-cemetery/src/fields/KeyNumbersCapital/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import { useState, useEffect } from 'react'
 import {
   Box,
   GridColumn,
