+++ libs/application/templates/inao/financial-statement-cemetery/src/utils/helpers.ts
@@ -1,49 +1,13 @@
 import { ExternalData, FormValue } from '@island.is/application/types'
-import { getValueViaPath } from '@island.is/application/core'
 import { BOARDMEMEBER, CARETAKER } from './constants'
 import { FinancialStatementCemetery } from '../lib/dataSchema'
 import getYear from 'date-fns/getYear'
 import subYears from 'date-fns/subYears'
 import { BoardMember, Config, FSIUSERTYPE } from '../types/types'
-import { TOTAL } from '../../../shared/utils/constants'
-
-export const getTotal = (values: Record<string, string>, key: string) => {
-  if (!values[key]) {
-    return 0
-  }
-  const total = Object.entries(values[key])
-    .filter(([k, v]) => k !== TOTAL && !isNaN(Number(v)))
-    .map(([_k, v]) => Number(v))
-    .reduce((prev, current) => {
-      return (prev += current)
-    }, 0)
-  return total
-}
-
-export const currencyStringToNumber = (str: string) => {
-  if (!str) {
-    return str
-  }
-  const cleanString = str.replace(/[,\s]+|[.\s]+/g, '')
-  return parseInt(cleanString, 10)
-}
-
-export const getCurrentUserType = (
-  answers: FormValue,
-  externalData: ExternalData,
-) => {
-  const fakeUserType: FSIUSERTYPE | undefined = getValueViaPath(
-    answers,
-    'fakeData.options',
-  )
-
-  const currentUserType: FSIUSERTYPE | undefined = getValueViaPath(
-    externalData,
-    'getUserType.data.value',
-  )
-
-  return fakeUserType ? fakeUserType : currentUserType
-}
+import {
+  getCurrentUserType,
+  currencyStringToNumber,
+} from '../../../shared/utils/helpers'
 
 export const getBoardmembersAndCaretakers = (members: Array<BoardMember>) => {
   const careTakers = members
