+++ libs/application/templates/inao/financial-statement-individual-election/src/fields/ElectionStatement/index.tsx
@@ -8,7 +8,7 @@ import { useFormContext } from 'react-hook-form'
 import { FinancialStatementIndividualElection } from '../../lib/utils/dataSchema'
 import { ELECTIONLIMIT } from '../../lib/utils/constants'
 import { m } from '../../lib/utils/messages'
-import { formatNumber } from '../../lib/utils/helpers'
+import { formatNumber } from '../../../../shared/utils/helpers'
 import { BottomBar } from '../../../../shared/components/BottomBar'
 import { GREATER } from '../../../../shared/utils/constants'
