+++ libs/application/templates/inao/financial-statement-individual-election/src/fields/Overview/index.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react'
 import { DefaultEvents, FieldBaseProps } from '@island.is/application/types'
-
 import {
   AlertBanner,
   Box,
