+++ libs/application/templates/inao/financial-statement-political-party/src/fields/ElectionStatement/ElectionStatement.tsx
@@ -8,7 +8,7 @@ import { useSubmitApplication } from '../../hooks/useSubmitApplication'
 import { useFormContext } from 'react-hook-form'
 import { FinancialStatementPoliticalParty } from '../../lib/dataSchema'
 import { ELECTIONLIMIT } from '../../utils/constants'
-import { formatNumber } from '../../utils/helpers'
+import { formatNumber } from '../../../../shared/utils/helpers'
 import { BottomBar } from '../../../../shared/components/BottomBar'
 import { GREATER } from '../../../../shared/utils/constants'
