+++ libs/application/templates/inao/financial-statement-political-party/src/fields/PowerOfAttorneyFields/PowerOfAttorneyFields.tsx
@@ -11,7 +11,6 @@ import {
 import { InputController } from '@island.is/shared/form-fields'
 import { useLocale } from '@island.is/localization'
 import { IdentityInput, Query } from '@island.is/api/schema'
-
 import { m } from '../../lib/messages'
 import { ABOUTIDS } from '../../../../shared/utils/constants'
 import { FieldBaseProps } from '@island.is/application/types'
