+++ libs/application/templates/inao/financial-statement-political-party/src/hooks/useTotals.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
-import { getTotal } from '../utils/helpers'
+import { getTotal } from '../../../shared/utils/helpers'
 
 export const useTotals = (key: string): [() => void, number] => {
   const [total, setTotal] = useState(0)
