+++ libs/application/templates/inao/financial-statement-political-party/src/utils/helpers.ts
@@ -1,7 +1,6 @@
 import { Config } from '../types/types'
 import subYears from 'date-fns/subYears'
 import getYear from 'date-fns/getYear'
-import { TOTAL } from '../../../shared/utils/constants'
 
 export const getConfigInfoForKey = (config: Config[], configKey: string) => {
   return config?.filter((config: Config) => config.key === configKey)[0].value
@@ -22,20 +21,3 @@ export const possibleOperatingYears = (
     })
   return operationYears
 }
-
-export const getTotal = (values: Record<string, string>, key: string) => {
-  if (!values[key]) {
-    return 0
-  }
-  const total = Object.entries(values[key])
-    .filter(([k, v]) => k !== TOTAL && !isNaN(Number(v)))
-    .map(([_k, v]) => Number(v))
-    .reduce((prev, current) => {
-      return (prev += current)
-    }, 0)
-  return total
-}
-
-export const formatNumber = (num: number) => num.toLocaleString('de-DE')
-
-export const checkIfNegative = (inputNumber: string) => Number(inputNumber) >= 0
