+++ libs/application/templates/inao/shared/utils/enums.ts
@@ -1,0 +1,5 @@
+export enum FSIUSERTYPE {
+  INDIVIDUAL = 150000000,
+  PARTY = 150000001,
+  CEMETRY = 150000002,
+}
